+++ CHANGES.md
@@ -1,3 +1,13 @@
+### Breaking changes
+- `BackendCommands` and `AsyncBackendCommands` are no longer decorated with `omit_exception`.
+- added `omit_exception_async` to decorate async operations, instead of using `omit_exception` for both sync and async.
+- `omit_exception` no longer supports async functions and generators.
+- added `DecoratedBackendCommands` and `DecoratedAsyncBackendCommands` as commands decorated with `omit_exception` and `omit_exception_async`.
+- `AsyncValkeyCache` and `ValkeyCache` no longer inherit from `BackendCommands` and `AsyncBackendCommands`, they inherit from `DecoratedBackendCommands` and `DecoratedAsyncBackendCommands` instead.
+
+### improvement
+- removed the undecorator loop from cluster client
+
 Version 0.3.2
 -------------